+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ## Current
 - :rocket: added capability to connect via CDP
+- :rocket: added I wait until current url step
+- :rocket: added I wait until page title step
 
 ## 0.0.6
 - :beetle: fixed issue in After hook if browser is not started